Refactor: 페이징 offset 타입 long으로 변경
 테스트를 통해, 트러블 슈팅을 하던 중 offset의 값이 정수의 최대값보다 커질 때, overflow가 발생하는 것을 발견하여,
long 타입으로 변경

+++ src/main/java/com/project/doubleshop/web/config/support/SimpleOffsetPageableHandlerMethodResolver.java
@@ -42,7 +42,7 @@ public class SimpleOffsetPageableHandlerMethodResolver implements HandlerMethodA
 			return null;
 		}
 
-		int offset = hasText(offsetString) ? parseAndApplyBoundaries(offsetString, Integer.MAX_VALUE) : simpleOffsetPageRequest.offset();
+		long offset = hasText(offsetString) ? parseAndApplyBoundaries(offsetString, Integer.MAX_VALUE) : simpleOffsetPageRequest.offset();
 		int limit = hasText(limitString) ? parseAndApplyBoundaries(limitString, DEFAULT_LIMIT_MAX_SIZE) : simpleOffsetPageRequest.limit();
 
 		limit = limit < 1 ? simpleOffsetPageRequest.limit() : limit;
